+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New getAppMetaInfos function with diskCache fallback
+- New getAppFiles with fallback strategies
 
 ## [3.55.4-beta] - 2019-10-09