+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - New getAppMetaInfos function with diskCache fallback
+- Add MineWinsConflict resolver logic to VBase client
 - New getAppFiles with fallback strategies
 
 ## [3.55.4-beta] - 2019-10-09
@@ -15,6 +16,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [3.55.3] - 2019-10-09
 ### Fixed
 - Messages Save v2 interface
+
+
+## [3.55.2] - 2019-10-01
 
 ### Fixed
 - Warnings logged as errors